requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coverage==7.3.0
Django==4.2.4
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
pip~=21.2.4
wheel~=0.37.0
cryptography~=36.0.0
lxml~=4.9.1
tomli~=2.0.1
colorama~=0.4.3
psycopg2~=2.9.2
PyYAML~=5.3.1
numpy~=1.23.5
setuptools~=68.0.0
future~=0.18.2